backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.6
firebase-admin==6.5.0
pydantic==2.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
emails==0.6
jinja2==3.1.2
httpx==0.27.0
email-validator==2.2.0 